Flights Delay Prediction-Random Forest Regressor
This exercise comprises of a set of data about airline flights to explore and analyze. The assignment is open-ended and unguided. 

The following data tasks are performed:

Data Exploration
What data are you given?
What do the data represent?
What relationships exist within the data?
What problems exist in the data, if any? How will you confront these problems?

Analysis
Build model(s) fleshing out some relationships and providing insights. These can be whatever you think is interesting. We are not concerned with what you do so much as how you do it and why.

Communication
As you work with the data, write out the steps you are taking and justify the decisions you make. This does not have to be done meticulously, but your work should be easy to follow by someone lightly reading your final report.
Any findings or analysis you do should be accompanied by an explanation of the findings.