Lead Scoring - Multiclass Logistic Regression
This commit uses the following dataset from the Airbnb challenge to predict booking destinations on Kaggle. The objective of this commit is to use the data to accomplish lead scoring using the test set (considering the test data as potential leads that haven't booked on the Airbnb site yet. The goal is to predict the non-US destination bookings with probabilities that can be used as a means to score the leads.

In this challenge, you are given a list of users along with their demographics, web session records, and some summary statistics. You are asked to predict which country a new user's first booking destination will be. All the users in this dataset are from the USA.

There are 12 possible outcomes of the destination country: 'US', 'FR', 'CA', 'GB', 'ES', 'IT', 'PT', 'NL','DE', 'AU', 'NDF' (no destination found), and 'other'. Please note that 'NDF' is different from 'other' because 'other' means there was a booking, but is to a country not included in the list, while 'NDF' means there wasn't a booking.

The training and test sets are split by dates. In the test set, you will predict all the new users with first activities after 7/1/2014 (note: this is updated on 12/5/15 when the competition restarted). In the sessions dataset, the data only dates back to 1/1/2014, while the users dataset dates back to 2010.