Bayesian Optimization_hyper parameter tuning
As the title describes, this commit has used bayesian optimization for tuning the hyper parameters and logistic regression has outperformed the random forest model. Although MLP classifier did very good in terms of performance compared to all other classifiers, tuning the MLP classifier is an whole other story. So not trying for now.